vmm: fix narrowing conversion warnings (19 violations)
- getphysicaladdr.c, unmappage.c: page index locals changed to u_int32_t
- vmm_memory.c: freePage/adjustCowCounter indices changed to u_int32_t,
  dead < 0 guards removed (unsigned / unsigned is always >= 0)
- paging.c: vmm_remapPage return type int→u_int32_t (returns a physical
  address used by callers as a non-null success indicator);
  vmm_remapIOPage returns 0 (no callers use the address return value);
  sys_brk td_retval[0] assignments use explicit (int) cast
- vmm_mmap.c: td_retval[0] pointer→int via (int)(uintptr_t); fix
  (caddr_t)-1 comparison to use tmp directly instead of retval
- getfreevirtualpage.c: explicit (int) cast on PT_INDEX subtraction

+++ sys/vmm/getfreevirtualpage.c
@@ -202,7 +202,7 @@ keepMapping:
 		{
 			map_from = start_page;
 		}
-		counter += PT_ENTRIES - PT_INDEX(start_page);
+		counter += (int)(PT_ENTRIES - PT_INDEX(start_page));
 		if (counter >= count)
 		{
 			goto gotRange;
